backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
